init split cluster files
