+++ paddle/cinn/frontend/cluster_ops/cluster_ops.h
@@ -1,0 +1,21 @@
+// Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+namespace cinn::api {
+
+ClusteringResult ClusterOps(
+    const std::vector<const pir::Operation*>& ops,
+    const std::shared_ptr<ShardableAxesProvider>& shardable_axes_provider,
+    const std::shared_ptr<ClusteringPolicy>& clustering_policy);
+}
